test(api): extract shared fake-D1 test harnesses for file metadata
Consolidate three copy-pasted node:sqlite fake-D1 harnesses
(file-metadata-sqlite/galleries-sqlite) into apps/api/test/helpers/sqlite-d1.ts
parameterized by migration path(s) and optional pragmas, and consolidate the
three near-identical hand-rolled file_metadata D1 mocks in routes-files.test.ts,
routes-public-files.test.ts, and usage-fake-d1.ts into a shared
FileMetadataTable helper that each suite's fake D1 tries first, falling back
to its own auth_tokens/workspace_usage logic. No behavior change.
